ARGO-3097 Tighten up PUT and POST on the backend API

+++ poem/Poem/api/internal_views/aggregationprofiles.py
@@ -12,6 +12,8 @@ from rest_framework.authentication import SessionAuthentication
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
+from .utils import error_response
+
 
 class ListAggregations(APIView):
     authentication_classes = (SessionAuthentication,)
@@ -20,27 +22,64 @@ class ListAggregations(APIView):
         serializer = serializers.AggregationProfileSerializer(data=request.data)
 
         if serializer.is_valid():
-            serializer.save()
+            try:
+                if request.data['groupname']:
+                    groupaggr = poem_models.GroupOfAggregations.objects.get(
+                        name=request.data['groupname']
+                    )
 
-            groupaggr = poem_models.GroupOfAggregations.objects.get(
-                name=request.data['groupname']
-            )
-            aggr = poem_models.Aggregation.objects.get(
-                apiid=request.data['apiid']
-            )
-            groupaggr.aggregations.add(aggr)
+                else:
+                    return error_response(
+                        status_code=status.HTTP_400_BAD_REQUEST,
+                        detail='You must provide a group of aggregations.'
+                    )
 
-            data = {
-                'endpoint_group': request.data['endpoint_group'],
-                'metric_operation': request.data['metric_operation'],
-                'profile_operation': request.data['profile_operation'],
-                'metric_profile': request.data['metric_profile'],
-                'groups': json.loads(request.data['groups'])
-            }
+                userprofile = poem_models.UserProfile.objects.get(
+                    user=request.user
+                )
 
-            create_profile_history(aggr, data, request.user)
+            except poem_models.GroupOfAggregations.DoesNotExist:
+                return error_response(
+                    detail='Given group of aggregations does not exist.',
+                    status_code=status.HTTP_404_NOT_FOUND
+                )
 
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
+            except poem_models.UserProfile.DoesNotExist:
+                return error_response(
+                    details='No user profile for authenticated user.',
+                    status_code=status.HTTP_404_NOT_FOUND
+                )
+
+            else:
+                if groupaggr in userprofile.groupsofaggregations.all() or \
+                        request.user.is_superuser:
+                    serializer.save()
+
+                    aggr = poem_models.Aggregation.objects.get(
+                        apiid=request.data['apiid']
+                    )
+                    groupaggr.aggregations.add(aggr)
+
+                    data = {
+                        'endpoint_group': request.data['endpoint_group'],
+                        'metric_operation': request.data['metric_operation'],
+                        'profile_operation': request.data['profile_operation'],
+                        'metric_profile': request.data['metric_profile'],
+                        'groups': json.loads(request.data['groups'])
+                    }
+
+                    create_profile_history(aggr, data, request.user)
+
+                    return Response(
+                        serializer.data, status=status.HTTP_201_CREATED
+                    )
+
+                else:
+                    return error_response(
+                        detail='You do not have permission to add resources to '
+                               'the given group.',
+                        status_code=status.HTTP_401_UNAUTHORIZED
+                    )
 
         else:
             details = []
@@ -49,40 +88,118 @@ class ListAggregations(APIView):
                     '{}: {}'.format(error, serializer.errors[error][0])
                 )
 
-            return Response(
-                {'detail': ' '.join(details)},
-                status=status.HTTP_400_BAD_REQUEST
+            return error_response(
+                detail=' '.join(details),
+                status_code=status.HTTP_400_BAD_REQUEST
             )
 
     def put(self, request):
         if request.data['apiid']:
-            aggr = poem_models.Aggregation.objects.get(
-                apiid=request.data['apiid']
-            )
-            aggr.groupname = request.data['groupname']
-            aggr.save()
+            try:
+                aggr = poem_models.Aggregation.objects.get(
+                    apiid=request.data['apiid']
+                )
 
-            groupaggr = poem_models.GroupOfAggregations.objects.get(
-                name=request.data['groupname']
-            )
-            groupaggr.aggregations.add(aggr)
+                init_groupaggr = None
+                if aggr.groupname and not request.user.is_superuser:
+                    try:
+                        init_groupaggr = \
+                            poem_models.GroupOfAggregations.objects.get(
+                                name=aggr.groupname
+                            )
 
-            data = {
-                'endpoint_group': request.data['endpoint_group'],
-                'metric_operation': request.data['metric_operation'],
-                'profile_operation': request.data['profile_operation'],
-                'metric_profile': request.data['metric_profile'],
-                'groups': json.loads(request.data['groups'])
-            }
+                    except poem_models.GroupOfAggregations.DoesNotExist:
+                        return error_response(
+                            detail="Initial profile's group of aggregations "
+                                   "does not exist.",
+                            status_code=status.HTTP_404_NOT_FOUND
+                        )
 
-            create_profile_history(aggr, data, request.user)
+                userprofile = poem_models.UserProfile.objects.get(
+                    user=request.user
+                )
 
-            return Response(status=status.HTTP_201_CREATED)
+            except poem_models.Aggregation.DoesNotExist:
+                return error_response(
+                    detail='Aggregation profile with given apiid does not '
+                           'exist.',
+                    status_code=status.HTTP_404_NOT_FOUND
+                )
+
+            except poem_models.UserProfile.DoesNotExist:
+                return error_response(
+                    detail='User profile for the given user does not exist.',
+                    status_code=status.HTTP_404_NOT_FOUND
+                )
+
+            else:
+                if init_groupaggr in userprofile.groupsofaggregations.all() or \
+                        request.user.is_superuser:
+                    try:
+                        if request.data['groupname']:
+                            groupaggr = \
+                                poem_models.GroupOfAggregations.objects.get(
+                                    name=request.data['groupname']
+                                )
+
+                        else:
+                            return error_response(
+                                detail='Please provide group of aggregations.',
+                                status_code=status.HTTP_400_BAD_REQUEST
+                            )
+
+                    except poem_models.GroupOfAggregations.DoesNotExist:
+                        return error_response(
+                            status_code=status.HTTP_404_NOT_FOUND,
+                            detail='Given group of aggregations does not exist.'
+                        )
+
+                    else:
+                        if groupaggr in userprofile.groupsofaggregations.all() \
+                                or request.user.is_superuser:
+                            aggr.groupname = request.data['groupname']
+                            aggr.save()
+
+                            groupaggr.aggregations.add(aggr)
+
+                            data = {
+                                'endpoint_group': request.data[
+                                    'endpoint_group'
+                                ],
+                                'metric_operation': request.data[
+                                    'metric_operation'
+                                ],
+                                'profile_operation': request.data[
+                                    'profile_operation'
+                                ],
+                                'metric_profile': request.data[
+                                    'metric_profile'
+                                ],
+                                'groups': json.loads(request.data['groups'])
+                            }
+
+                            create_profile_history(aggr, data, request.user)
+
+                            return Response(status=status.HTTP_201_CREATED)
+
+                        else:
+                            return error_response(
+                                detail='You do not have permission to change '
+                                       'resources in the given group.',
+                                status_code=status.HTTP_401_UNAUTHORIZED
+                            )
+
+                else:
+                    return error_response(
+                        detail='You do not have permission to change resources '
+                               'in the given group.',
+                        status_code=status.HTTP_401_UNAUTHORIZED
+                    )
 
         else:
-            return Response(
-                {'detail': 'Apiid field undefined!'},
-                status=status.HTTP_400_BAD_REQUEST
+            return error_response(
+                detail='Apiid field undefined!',
+                status_code=status.HTTP_400_BAD_REQUEST
             )
 
     def get(self, request, aggregation_name=None):
@@ -99,11 +216,14 @@ class ListAggregations(APIView):
                 return Response(serializer.data)
 
             except poem_models.Aggregation.DoesNotExist:
-                raise NotFound(status=404,
-                               detail='Aggregation not found')
+                raise NotFound(
+                    status=404, detail='Aggregation not found'
+                )
 
         else:
-            aggregations = poem_models.Aggregation.objects.all().order_by('name')
+            aggregations = poem_models.Aggregation.objects.all().order_by(
+                'name'
+            )
             serializer = serializers.AggregationProfileSerializer(
                 aggregations, many=True
             )
@@ -115,21 +235,55 @@ class ListAggregations(APIView):
                 aggregation = poem_models.Aggregation.objects.get(
                     apiid=aggregation_name
                 )
-                poem_models.TenantHistory.objects.filter(
-                    object_id=aggregation.id,
-                    content_type=ContentType.objects.get_for_model(aggregation)
-                ).delete()
-                aggregation.delete()
-                return Response(status=status.HTTP_204_NO_CONTENT)
+                groupname = None
+                if aggregation.groupname:
+                    try:
+                        groupname = poem_models.GroupOfAggregations.objects.get(
+                            name=aggregation.groupname
+                        )
+
+                    except poem_models.GroupOfAggregations.DoesNotExist:
+                        pass
+
+                userprofile = poem_models.UserProfile.objects.get(
+                    user=request.user
+                )
+
+                if (
+                        groupname and groupname in
+                        userprofile.groupsofaggregations.all()
+                ) or request.user.is_superuser:
+                    poem_models.TenantHistory.objects.filter(
+                        object_id=aggregation.id,
+                        content_type=ContentType.objects.get_for_model(
+                            aggregation
+                        )
+                    ).delete()
+                    aggregation.delete()
+                    return Response(status=status.HTTP_204_NO_CONTENT)
+
+                else:
+                    return error_response(
+                        detail='You do not have group permission to delete this'
+                               ' aggregation profile.',
+                        status_code=status.HTTP_401_UNAUTHORIZED
+                    )
 
             except poem_models.Aggregation.DoesNotExist:
-                raise NotFound(status=404,
-                               detail='Aggregation not found')
+                raise NotFound(
+                    status=404, detail='Aggregation not found'
+                )
+
+            except poem_models.UserProfile.DoesNotExist:
+                return error_response(
+                    detail='No user profile for authenticated user.',
+                    status_code=status.HTTP_404_NOT_FOUND
+                )
 
         else:
-            return Response(
-                {'detail': 'Aggregation profile not specified!'},
-                status=status.HTTP_400_BAD_REQUEST
+            return error_response(
+                detail='Aggregation profile not specified!',
+                status_code=status.HTTP_400_BAD_REQUEST
             )
